feat: ellipsizeMode e2e tests
