+++ apps/example-web/src/RouteSelector.tsx
@@ -5,6 +5,7 @@ import { TestSetSelection } from './testScreens/TestSetSelection';
 import { VisualRegression } from './testScreens/VisualRegression';
 import { TestSubmitProps } from './testScreens/TestSubmitProps';
 import { TestEnrichedText } from './testScreens/TestEnrichedText';
+import { TestEllipsize } from './testScreens/TestEllipsize';
 import { useEffect, useState } from 'react';
 
 export default function RouteSelector() {
@@ -43,6 +44,10 @@ export default function RouteSelector() {
 
   if (path === '/test-enriched-text') {
     return <TestEnrichedText />;
+  }
+
+  if (path === '/test-ellipsize') {
+    return <TestEllipsize />;
   }
 
   return <App />;
